Split Calldata into multiple files - 1 class per file
