+++ packages/utils/src/abi_encoder/index.ts
@@ -1,2 +1,2 @@
-export { EncodingRules, DecodingRules } from './calldata';
+export { EncodingRules, DecodingRules } from './utils/rules';
 export * from './evm_data_type_factory';
