+++ packages/utils/src/abi_encoder/utils/rules.ts
@@ -1,0 +1,8 @@
+export interface DecodingRules {
+    structsAsObjects: boolean;
+}
+
+export interface EncodingRules {
+    optimize?: boolean;
+    annotate?: boolean;
+}
